refactor: extract command helper
